fix(api): gate parentLinkDates behind translate-access permission check
- Store parentLinkDates as a true/absent flag instead of mirroring the
  parent's raw linkDates value, so downstream code can't rely on it
  being present without permission.
- Extract hasTranslateAccessToAllTranslations into its own module for
  reuse.
- Update dev-*.sh worktree paths/ports and .gitignore entry for the
  new worktree.
