+++ .gitignore
@@ -8,4 +8,4 @@ stats.html
 
 # Throwaway SSG proof-of-concept (kept locally for reference, not part of the repo)
 vite-ssg_poc/
-.worktree-ports
+.worktree-ports.env
