--- .worktree-ports.env
@@ -1,6 +1,0 @@
-# Generated by the Orca worktree Setup Script — safe to delete/regenerate.
-LUMINARY_WORKTREE_SLOT=3
-API_PORT=3030
-API_TEST_PORT=3032
-APP_PORT=4204
-CMS_PORT=4205
