+++ dev-api.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env bash
 set -euo pipefail
-cd "/Users/dirk/orca/workspaces/luminary/1672-app-research-and-implement-vue-ssr-for-ssgisr/api"
+cd "/Users/dirk/orca/workspaces/luminary/278-cms-add-toggle-to-link-dates-on-content-documents-belonging-to-the-same-parent/api"
 npm run start:dev
