+++ dev-cms.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env bash
 set -euo pipefail
-cd "/Users/dirk/orca/workspaces/luminary/1672-app-research-and-implement-vue-ssr-for-ssgisr/cms"
-npm run dev -- --port 4205
+cd "/Users/dirk/orca/workspaces/luminary/278-cms-add-toggle-to-link-dates-on-content-documents-belonging-to-the-same-parent/cms"
+npm run dev -- --port 4265
